Compilation progress should display much better now. Rather than overriding run, developers should override preDo and postDo instead.
git-svn-id: https://svn.code.sf.net/p/gjava/code@1362 0bf6c8e7-452b-0410-8c87-9aaedaf074f6

+++ G-CSharp/src/org/gcreator/compilers/GCS/libGCS/Scripting.Math.cs
@@ -174,7 +174,7 @@ namespace org.gcreator.Scripting
         {
             int max = (int) x.getDouble();
             System.Random r = new System.Random();
-            return new Double(r.Next(max));
+            return new Integer(r.Next(max));
         }
 
         public static Object choose(params Object[] args)
